[FIX] web: many2many attrs computation
attrs with domains based on a many2many fields
do not worked properly. Indeed, to set invisible based on a domain
[('field_ids', '=', False)],
field_ids being a many2many, you actually had to write
[('field_ids', '=', [(6, False, [])])]
Which is not really user friendly.

Using field.get('value') instead of field.get_value has the advantage
to always returns the ids list, for *2many fields, instead of
a list of commands.

opw-629358