[FIX] fields: insert xml_id for each model field re-definition
When uninstalling a module,
a field is unlinked only if there
are no more external identifiers for this field.

Before this revision, a field external identifier
was created only upon the field creation,
while it should be created as well if the field
is already there (so do not need
to be created, only updated according
to the new field attributes) if
the fields is declared in another module.

Otherwise, if both modules `A` & `B` declare
a field `X` on model `Y`, and both modules are
installed, but module `A` is installed first,
only the external identifier of module `A`
for this field will be created.

This is problematic if you uninstall module
`A` without uninstalling module `B`
(when there is no dependence between each other),
as it will remove the external identifier of
module `A`, which had the only external identifier
for this field `X`. As there is no more
any external identifier for this field, it will
be dropped, while module `B` still needs it.

opw-647968