[FIX] website_forum: show to the end user that he should be logged in
Show to the end user that he should be logged in to use action where karma
is required.

Before this commit, a user not logged see that x karma is required to perfom
the action.

After this commit, a user not logged who click on an action which require karma
will get a message "Sorry you must be logged to perform this action"

This commit closes #12731