[FIX] db.py: list_dbs: PGUSER handling
PGUSER is an environment variable used by lipq and thus psycopg to
connect to the databse and should be used if not --db_user argument has
been sent to Odoo. However, the list_dbs method did not fallback on this
environment variable if a --db_user was not present but instead chose
the user that executed the Odoo process. So you created the db with an
user and try to list it with another.