[IMP] product: add weight widget
The weight widget allows you to display a database float value (stored in
the default uom (kg)) in the unit of measure defined in the general settings.

Both form and kanban widgets have been added, they both go by the name weight.
E.g. (in a form or kanban view):
   <field name="my_weight_field" widget="weight">