[IMP] report_intrastat: add weight conversion
Conversion to the report.intrastat report, it will take the UoM of the
invoice's company.

It could be arguable whether to take the product or invoice's company,
or to make no conversion at all...

I'm not sure about this one. But atlease the report groups by company, so
no values from different units of measures are added.

Maybe we shouldn't display the weight, as no conversion with the widgets can
be made there.