[FIX] website_sale: fix amount format
'/website/translations' is only called when user have right to edit page.
So a standard user didn't call it and l10n is not initialized.

After an update, now we format the amount with the l10n value.
To stay retro compatible, if l10n is not initialized (value = [])
we use [] as grouping as 'fallback value'.

To fix decimal precision you need to update the template product_price.

To fix the grouping, you need to update the website.layout

To fix the decimal separator, (and previous fix), you just need to pull the JS

This commit is related to #<1103, #<11553, #<14772, #<14874, ...
And fix the previous fix odoo/odoo@1f10ef8055f6c661fc6407a8ce4593dfc30452b4