[FIX] Propagation field doen't update in stock settings view.
The bug happens when you set the "Minimum days to trigger a propagation of date change in pushed/pull flows" field in the stock settings view.
If you apply your changes and come back to the settings view, the previous field display 0 as current value.

The propagation_minimum_delta field was properly updated in the model when we save the settings. The only problem come from the view because the company_id field was not present in the view and so the related field propagation couldn't retrieve his data.

This commit fix the problem by updating the related field propagation_minimum_delta (not mandatory) and adding the invisible field company_id in the stock settings xml form view.

This bug was resolved in master branch after new settings commit 31cb715787930ea346db80336b1b95a9fcb793b6 but it remains from odoo 9 to saas-14.
------------------------------------------------------------

There was an other bug that displays the line "Minimum days to trigger a propagation of date change in pushed/pull flows" when The checkbox "No automatic routing of preoducts" was checked.

The problem come from the condition for invisibility {'invisible': [('group_stock_adv_location', '=', 0)]}. In the field declaration the 0 was declared as an integer and not a string this implies the 0 to be treated as False and break the condition.

To correct this problem, the condition now check for a False instead of 0.