[FIX] sale, purchase: better info on qty update in purchase/sale order line
When the quantities on a purchase or sale line decrease,
the amount of the receipt or delivery order is not adapted.
Also, when the quantities increase, it is good to follow it up.

That is why, when we change some quantity, a message is
posted in the chatter of the purchase/sale to indicate about
the change.