[FIX] account: when changing settings, if currency has not change, don't update it
this increases performance, reason is because when saving a res.config.setting, even if a field has not changed, they will be updated and this might trigger recomputation, and so on... Problem here is that when writing a currency_id to the company, all moves are impacted due to the company_currency_id which is a related stored field. And saving the settings in accounting can be pretty slow if we have lots of move. This fix reduce the time needed to save settings by not writing currency_id on the company if the value has not changed.