[FIX] mass_mailing: get recipients as sudo write_uid
Before this the list of recipients of a mass mailing was computed by the
root user and always with language en_US.

With this commit, the last modifier (using "Send to All" qualify) of the
mass mailing is used to compute the recipients list (or if not present
the mailing creator and lastly the root user as it was before).

Pros:

- records rule when writing the mass mailing and selecting users are
  applied when sending the mail,

- if the root user has a different language than the mass mailing sender
  a domain dependent on language will no longer have different results

Cons:

if someone else write on the mass mailing between click on "Send to All"
and the end of mail sending the recipients list could be changed (if the
user had other language or record rule).

opw-709950