[IMP] point_of_sale: trust client totals
The amounts sent by the point of sale should be trusted instead of recomputing
it server side.
The previous compution could introduce differences in amounts.

Turn the methods into onchange to still be possible to do refund but disable
edition from the backend.

The pos.orders should only be created from the pos interface.

change amount_paid computation with the one server side

The way amount_paid is computed was not coherant

Adapt tests to use a computed amount_total and amount_paid