[IMP] sale/purchase/report: specify reporting UoMs
Reports will go fetch the database values of weights and volumes,
stored in Kg an Liters. Specify it in the label otherwise the user
can't know.

Other option would've been to convert into the current company's UoM
but that would imply not displaying the UoM, and making the user assume
it's in his company's UoM.