[FIX] web: grouped kanban view was not full-height
In the community version, the grouped kanban view was not full-height
anymore. This was due to the useless controller div wrapping the view
(the renderer associated DOM).

Controller should not be Widget instances, the rendering part should be
entirely left to the renderer. As this would require some refactoring
for some views, this commit is making a temporary fix for basic
controllers: at the end of the `start` current behavior, we make the
controller `$el` element to be equal to the renderer `$el` element
thanks to the `replaceElement` function to remove the useless wrapping
<div/> element the controller created.