[FIX] account: fix computing of tax included in price
Before this commit, there was no way to compute two taxes that apply on one same amount
The formula used for this use case was always:
BasePrice = TotalPrice / PRODUCT( (1 + tax.rate) for tax in TAXES)
This formula is fine if one tax computes the result for the next one but NOT if both taxes apply on the same base amount

In this latter case the formula is:
BasePrice = TotalPrice / 1 + SUM(tax.rate for tax in taxes)

For example: Total price = 126, composed by base price, TVA tax 21%, TVA tax 5% (all price included)
We compute base price = total price / 1.26 = 100

This commit fixes this. Obviously we need to pass the rate of a tax to the next one to compute the sum of rates, which is done in two main passes.
It is strongly advised in this commit to make use of this feature as a group of taxes UNIFORMLY formed by taxes having these caracteristics:
- Included in Price = True
- Affect subsequent Taxes = False
- Computation method = percent

OPW 741105