[FIX] fix bad computation of 'Won in opportunities' in crm dashboard
The filter on activity_date_deadline has been added during a fix, see 28c158cd130ce491b256e65b8f48c1f140fff24d
It's incorrect for the computation of the 'Won in opportunities' because it excludes opportunities not having
activity_ids.

Furthermore, the way to compute the 'won' amount is not equals to the filters when clicking on the button.
Then, we need to keep only opportunities standing in a kanban column having a probability of 100%.

-opw 742489