[FIX] web_tour: tips were overlapping the chat windows
Before this commit, the tour tips were overlapping the chat windows.
This was due to the fact they always had a z-index of 1051 wherever
they were. If a tip was placed in a location where none of its
ancestors created a new stacking context, the closed tip was then
positioned above everything else, including the chat windows. A
simple solution is made by this commit: only set the z-index when
the tooltip is opened (and thus in the body), as in that case, we
want it to be above everything else. This is however not entirely
correct. Even closed, we might want the tips to be above elements
they are suppposed to be overlapping.