[ADD] base: add modules_dependent property
Little trick to allow having fields invisible depending of
the installation of some modules.

Having fields with modules_dependent='mod1,mod2,...' means the field
will be invisible on views when mod1,mod2,... are not all installed.

It's useful when dealing with account/account_accountant to have some
fields invisible when account is installed but not account_accountant.