[FIX] web: Display external identifiers
When checking an external identifier from the technical settings (debug
mode), clicking on an external identifier entry would give an error and
would not display the external identifier entry at all.

This was because field_utils.format.many2one was being passed a string
as its value parameter, when the method requires either an Array, an
Object (record) or false.

This commit fixes the bug by passing in this.record which is an object,
instead this.value which is a string