[FIX] test_main_flows: wait save before editing many2one
Since @c300e5ed, the many2one line is redrawn in case of editing.

In the main_tour, we tried to edit a many2one in a one2many.
A modal was displayed and we made some changes. Then we clicked on
the save button and we direclty clicked to add a new record.
The problem was that we were trying to modify this field
a second time without waiting the save.

The following message was displayed:
"The record has been modified, your changes will be discarded.
Are you sure you want to ?"

In fact, the main_tour clicked to add a second row
before the modal is closed (this modal is closed after saving).

So now, we wait that this modal is close before adding a second
record.