[FIX] web_editor: enter in br very buggy
There is an odd behavior since c227b8fd88, if for example we press:

XYZ[shift+enter]
[shift+enter]

we get as result:

YZ.

instead of:

XYZ[newline]
[newline]

This is because the code of `$.summernote.pluginEvents.visible` when the
target location is an 'dom.isForbiddenNode' has a logic which is broken
by other code which change the range between the setimeout.

Thus we expect to get a `.` text node but instead get a element node for
which the code is completely wrong and do rubish.

fixes #17719