[FIX] base_import: handle parsing of relational fields chains
It is possible to import a value for a subfield of a relational
field of the model by describing the relational field chain using
slash '/' characters.
Example: product_id/seller_ids/price

However, those values were completely ignored by the parsing
phase of the import feature. This commit fixes that.

This commit also changes the way the parsing function looks for
fields. Rather than looking at all the fields of the model and
see if we are importing these ones, we know only look for the
fields that we are actually importing.