[FIX]: website, website_hr: fix error when user rename the page of /aboutus when website_hr is installed. The module was doing an xpath on the ir.ui.view 'aboutus' adding a loop to show employees. The employee variable used in the loop in xml was then set in a controller intercepting the @route /aboutus and rendering the ir.ui.view with a value 'employee'.
But now with website.page, if the user change the page of /aboutus, it won't trigger the controller adding the employee variable and it will throw an error when website.page will render the ir.ui.view without the employee variable.
So we should set it in the xpath to keep the variable being set without going in the controller.