[FIX] stock: modify product type with open move lines
When the user change the type of a product to stockable. It could
exists some move line with a quantity reserved for this product.
Thus when the user will confirm the picking, he will have an issue
because the system try to remove reservation quants that are not
present.

Thic commit add a constraint that check if there is open move lines
for this product.