[FIX] web: expose additional options to python actions
Some options were added to the action manager (in the do_action method),
as options, not in the action description.  This means that these
options are not accessible for actions coming from the server.  In this
commit, we also use the server value to determine what options we need
to apply.