[FIX] stock: merge move's origin in picking origin
In a procurement group case move are merged in a single
picking. However picking origin only contains the origin
of first included move(s).

This commit modifies assign_picking in order to add the origin
of current move inside the origin of the picking if it does not
already exist