[FIX] web: refresh company image after dialog edit
There was several issue when editing a company data by clicking on the
logo (in odoo community):

- the logo was saved but never refreshed
- the form was not closed after saving
- the code to refresh the logo did not refresh the logo

This commit replace the old code executing an action by the opening of a
FormViewDialog to edit the company data and fix the three above points.

To force the reloading of the company logo, &_={nonce} is used in the
URI query instead of changing URI fragment because at least current
chromium doesn't seem to try to refresh an image based on such a change.