[IMP] website, website_sale: unify structure of HTML and better css/design for dashboard
Add margin to Sales title in dahsboard content.
Change the way the tooltip is left aligned.
Change structure of dashboard content for sales block to get the same HTML structure than Visits.
Change the condition of a t-if that would be always true if there is one order.
Don't override the hover color in table
Don't change the color when title of table is hovered.