[FIX] web: do not trigger `do_action` in many2one widget
Before this rev, the many2one widget was triggering the `do_action` itself.
As this event is directly managed by the webclient and not a controller,
the event was catched even if the field was in a form only rendered (for example
in Studio) without controller.

We fix this by handling this event on the controller (FieldManagerMixin in
this case) level.

A test case will be added in the Studio test suite.