[REF] purchase_requisition: PO creation from blanket order
If a there is blanket order in 'ongoing' state and linked with the purchase
order's vendor, we can choose associate it to the PO to retrieve the contrat
price and impact the ordered quantity.

If a purchase order is created from a blanket order, the vendor is
linked to the BO reference. So we forbid its edition once its choose.