[FIX] stock_account: in or ou move consideration
Only looking on the stock move to determine if a move should create
accounting entries is not enough: it is possible that sublocations
implement some strategies that impact the valuation.

We now look on the move line and if any of the move line is IN, the move
is IN, if any of the move line is OUT, the move is OUT, if there's a
mix, the move is invalid.