[IMP] web: FieldChar trim its value by default
With this commit, we change the behaviour of the web client with respect
to spaces in char fields.  Most of the time, starting and ending spaces
have no value, and worse, make the data not so reliable.

After this commit, field char will trim by default (so, if the user input a
char as ' abc ', the string 'abc' will be sent to the server instead).
Note that this only applies when the value of the field is changed.  If
someone open a form view, then switches to edit mode and save, nothing
will change.

This is the desired behavior most of the time.  However, in some rare
case, this is actually harmful.  For example, if we trim the
'decimal_point' field, it will not be possible to enter a space as
decimal separator.  In those cases, we introduce a new attribute 'trim',
which allow the developer to desactivate that feature.