[REF] tests: add a pico language to select tests
This commits introduce prefixes to select cases with tags.
'+' or no prefix means that the tests tagged with this tag are selected
for execution. '-' prefix means exclude the tests tagged with this tag.
Exclusion takes precedence over inclusion.
Also, by default, all tests without tags are considered tagged
'standard'. This means that when selecting tests with the 'test-tags'
parameter, if '-standard' is not specified, all tests without tags are
going to be executed.