[FIX] web_planner: use user language on frontend
The web planner specified no language for frontend so the website
planner on the frontend was never translated.

This commit add a specialized fallback for this use case so the planner
on the website is in the language of the current user.

A better fix might be to really have a context on the website (thus
making require('web_editor.base').get_context unecessary) but:

- this would be a bigger and more risky change,
- these part have changed between 9 and 10 and between 10 and 11

note: this is for 9.0 up to not including 11.0

opw-781727
closes #21004