[FIX] website...: have user language in context
Most locations of the website use web_editor.context to get sort of a
dummy context containing the language (and possibly version/website).

The web planner did not and in 9.0 up to saas-16 this has been solved by
a fix with [...].

In this change for 11.0, what is done is having this dummy context
inside the session so the context is gotten in the frontend as it was
in the backend.

opw-781727
closes #21035