[REF] tests, config: parse tags parameters only once
Before this commit, test-tags parameter was parsed for each test.
With this commit, it's parsed at the beginning only.
Also, there is now a function to parse the test-tag parameter
to could be extended later. It returns a tuple of two sets, the first
one being tags to select test and the second, tags to deselect tests.