[FIX] payment_authorize: some logging required
Authorize: International Provider of Misteries

It seems in my development/merge of the authorize.net API, I did not
stop and think about how hard it would be to actually debug this.
This is made even more difficult by the fact that authorize.net does
not (AFAICT) display payment errors on their backend. This means we
receive failure responses without any explicit message.