[FIX] base: do not unlink() `ir.property` records when setting fields
When several company-dependent fields are written, and the first one is set to
its default value, we delete the existing property record with `unlink()`,
which clears the record cache.  As a consequence, the other company-dependent
fields are set to their current value, because their new value has been lost.
In order to avoid the cache invalidation, use plain SQL instead of `unlink()`.

OPW 782080