[FIX] barcodes: manage several barcode widgets
related to rev 0d19ee989d1d890860adaf360de204eb543136a5

It was broken following the refactoring done in saas16, resulting in
barcodes being handled on all the displayed forms and it was problematic
in the case of a form dialog openned over a "fullscreen" form view.

This was particularily problematic for the scrap dialog in the
stock_barcode module.