[ADD] auth_token: token based authentication
This new module add the possibility for users with the 'Access Right'
group to create time-limited token that can then be shared via emails
(or through a simple link). The goal of this module is to allow third
parties to connect to the database for assistance purposes; e.g. to
debug or assess problems without having to share the normal credentials
of the user that will be impersonated.

In a nutshell, a new model `auth.token` is introduced; this model has
a time limit and is linked to a `res.users` record. When an
authentication attempt is made using the token as password, an override
of the `check_credentials` method of the `res.users` model checks the
provided token against a hash stored outside of the ORM's scope. Expired
tokens are collected by an `ir.cron` record that runs every day by
default.

This application also adds an entry in the Settings Dashboard to easily
create these tokens (as well as a view extension for the form view of
`res.users` records).