[IMP] website_rating_project: redesign project rating page
This commit improves and cleans the project rating page;
 - statistics of top 5 rated in last 15 days in right column of the page
 - move progress bar for tasks on right side of the page
 - show the detail of a rating when mouse hovering on the smiley

Since the page is refreshed, the statistics computation is done
with a single SQL request to extract all raw data to calculate
period and partner stats to display.

Thanks to @ror-odoo for the first implementation
and to @fda-odoo for its review.

Thanks too to @jem-odoo for its review