[IMP] web: fieldutils toggle_float widget
Define format and parse for toggle_float field widget,
so that it reuses the same as for float.

We need it in order to display the cell contents as float,
instead of float_time. We have to name it 'toggle_float'
so that we can put some extra behavior in timesheet grid,
which let us toggle the value by clicking on any cell.