[FIX] hr_expense: access error
Commit 957d57eaf7b0 added the auto-subscription of managers to an
expense at creation. However, this causes a conflict with the `ir.rule`
`mail_followers_read_write_own`. If an employee has a manager, an access
error is raised upon expense sheet creation.

Adding a `sudo` is not an option, because the chatter will log the admin
as the creator of the expense. Therefore, we are doing a workaround by
using an intermediary 'Draft' state.

opw-785416